+++ code/include/algorithms.hpp
@@ -29,51 +29,17 @@ bool parse_cpx_status(const CPXENVptr& cpxenv, const CPXLPptr& cpxlp, hplus::env
 // ===================== HEURISTICS ===================== //
 // ====================================================== //
 
-void find_heuristic(hplus::instance& inst, hplus::environment& env, const logger& log);
+void run_heur(hplus::instance& inst, hplus::environment& env, const logger& log);
 
 double find_hmax_goal(const hplus::instance& inst, const logger& log);
 
 // ====================================================== //
-// ======================== IMAI ======================== //
+// ======================== MODEL ======================= //
 // ====================================================== //
 
 void cpx_build_imai(CPXENVptr& cpxenv, CPXLPptr& cpxlp, const hplus::instance& inst, const hplus::environment& env, const logger& log,
                     bool relaxed = false);
 
-void cpx_post_warmstart_imai(CPXENVptr& cpxenv, CPXLPptr& cpxlp, const hplus::instance& inst, const hplus::environment& env, const logger& log);
-
-void store_imai_sol(const CPXENVptr& cpxenv, const CPXLPptr& cpxlp, hplus::instance& inst, const logger& log);
-
-// ====================================================== //
-// ======================= RANKOOH ====================== //
-// ====================================================== //
-
-void cpx_build_rankooh(CPXENVptr& cpxenv, CPXLPptr& cpxlp, const hplus::instance& inst, const hplus::environment& env, const logger& log);
-
-void cpx_post_warmstart_rankooh(CPXENVptr& cpxenv, CPXLPptr& cpxlp, const hplus::instance& inst, const hplus::environment& env, const logger& log);
-
-void store_rankooh_sol(const CPXENVptr& cpxenv, const CPXLPptr& cpxlp, hplus::instance& inst, const logger& log);
-
-// ====================================================== //
-// ==================== DYNAMIC SMALL =================== //
-// ====================================================== //
-
-void cpx_build_dynamic_small(CPXENVptr& cpxenv, CPXLPptr& cpxlp, const hplus::instance& inst, const hplus::environment& env, const logger& log);
-
-void cpx_post_warmstart_dynamic_small(CPXENVptr& cpxenv, CPXLPptr& cpxlp, const hplus::instance& inst, const hplus::environment& env,
-                                      const logger& log);
-
-void store_dynamic_small_sol(const CPXENVptr& cpxenv, const CPXLPptr& cpxlp, hplus::instance& inst, const logger& log);
-
-// ====================================================== //
-// ==================== DYNAMIC LARGE =================== //
-// ====================================================== //
-
-void cpx_build_dynamic_large(CPXENVptr& cpxenv, CPXLPptr& cpxlp, const hplus::instance& inst, const hplus::environment& env, const logger& log);
-
-void cpx_post_warmstart_dynamic_large(CPXENVptr& cpxenv, CPXLPptr& cpxlp, const hplus::instance& inst, const hplus::environment& env,
-                                      const logger& log);
-
-void store_dynamic_large_sol(const CPXENVptr& cpxenv, const CPXLPptr& cpxlp, hplus::instance& inst, const logger& log);
+void run_model(hplus::instance& inst, hplus::environment& env, hplus::statistics& stats, const logger& log);
 
 #endif /* ALGS_H */